fix(chat): consume 't' character to prevent ghost input on open     feat(king-for-four): implement server-side victory detection and     client Game Over screens     style(king-for-four): polish table with gradients, shadows, and     opponent card backs     feat(lobby): introduce interactive Main, Pause, Settings, and     Leaderboard menus     feat(firstparty): add binary-serialized Progress (AP, skins, crown     jewels) and Leaderboard systems     tweak(lobby): require ENTER to launch games from trigger zones     fix(lobby): prevent window auto-close on ESC, re-enable standard     window borders     feat(reseau): send full LobbyState on client join and interpolate     network movements"

+++ king-for-four/src/core/game.c
@@ -7,6 +7,7 @@
 
 #include "core/game.h"
 #include "core/card.h"
+#include "baseTypes.h"
 #include <stddef.h>
 #include <stdio.h>
 
@@ -15,7 +16,7 @@
  * [Discard Pile Color (0-3)][Played Card Color (0-4)]
  * 1 indicates a valid move, 0 indicates an invalid move.
  */
-int valid_color[4][5] = {
+s32 valid_color[4][5] = {
     {1, 0, 0, 0, 1},
     {0, 1, 0, 0, 1},
     {0, 0, 1, 0, 1},
@@ -26,14 +27,14 @@ int valid_color[4][5] = {
  * @brief Rule matrix for valid value moves.
  * [Discard Pile Value (0-14)][Played Card Value (0-14)]
  */
-int valid_value[15][15] = {0};
+s32 valid_value[15][15] = {0};
 
 /**
  * @brief Initializes the game logic, rules, and state.
  * @param g Pointer to the GameState to initialize.
  */
 void init_game_logic(GameState* g) {
-    for (int i = 0; i < 15; i++) valid_value[i][i] = 1;
+    for (s32 i = 0; i < 15; i++) valid_value[i][i] = 1;
 
     if (!g) return;
     g->current_player = 0;
@@ -44,7 +45,7 @@ void init_game_logic(GameState* g) {
     g->discard_pile = (Deck){NULL, 0};
     
     // INITIALISATION CRITIQUE DES MAINS
-    for (int i = 0; i < 4; i++) {
+    for (s32 i = 0; i < 4; i++) {
         g->players[i].hand.head = NULL;
         g->players[i].hand.size = 0;
     }
@@ -57,11 +58,11 @@ void init_game_logic(GameState* g) {
  * @param top The card on top of the discard pile.
  * @return 1 if the move is valid, 0 otherwise.
  */
-int is_move_valid(int active_color, Card played, Card top) {
-    int idx = (active_color >= 0 && active_color < 4) ? active_color : top.color;
+s32 is_move_valid(s32 active_color, Card played, Card top) {
+    s32 idx = (active_color >= 0 && active_color < 4) ? active_color : (s32)top.color;
     
-    int color_ok = (idx >= 0 && idx < 4) ? valid_color[idx][played.color] : 0;
-    int value_ok = valid_value[top.value][played.value];
+    s32 color_ok = (idx >= 0 && idx < 4) ? valid_color[idx][played.color] : 0;
+    s32 value_ok = valid_value[top.value][played.value];
 
     return color_ok || value_ok;
 }
@@ -73,15 +74,15 @@ int is_move_valid(int active_color, Card played, Card top) {
 void distribute_cards(GameState* g) {
     if (!g || g->num_players <= 0) return;
 
-    for (int i = 0; i < g->num_players; i++) {
-        for (int j = 0; j < 7; j++) draw_to_hand(&(g->players[i]), &(g->draw_pile));
+    for (s32 i = 0; i < g->num_players; i++) {
+        for (s32 j = 0; j < 7; j++) draw_to_hand(&(g->players[i]), &(g->draw_pile));
     }
 
     Card first = pop_card(&(g->draw_pile));
     push_card(&(g->discard_pile), first);
     
     /* Couleur forcée si la première carte est un Joker */
-    g->active_color = (first.color == 4) ? 0 : first.color;
+    g->active_color = (first.color == 4) ? 0 : (s32)first.color;
 }
 
 /**
@@ -91,12 +92,12 @@ void distribute_cards(GameState* g) {
  * @param cardIndex Index of the card in the player's hand.
  * @return 1 on success, 0 on failure.
  */
-int try_play_card(GameState *g, int playerIndex, int cardIndex) {
+s32 try_play_card(GameState *g, s32 playerIndex, s32 cardIndex) {
     Player* p = &g->players[playerIndex];
     
     // 1. Trouver la carte dans la main sans la retirer 
     Node* current = p->hand.head;
-    for (int i = 0; i < cardIndex; i++) {
+    for (s32 i = 0; i < cardIndex; i++) {
         if (current) current = current->next;
     }
     if (!current) return 0; // Erreur d'index
@@ -140,7 +141,7 @@ int try_play_card(GameState *g, int playerIndex, int cardIndex) {
  * @param playerIndex Index of the player drawing.
  * @return 1 on success, 0 if no cards are available.
  */
-int player_draw_card(GameState *g, int playerIndex) {
+s32 player_draw_card(GameState *g, s32 playerIndex) {
     Player* p = &g->players[playerIndex];
 
     // 1. Si la pioche est vide, on recycle le talon
